+++ lib/lighthouseProvider/backEnd/FlutterWebBluetoothBackEnd.dart
@@ -6,10 +6,10 @@ import 'package:flutter_web_bluetooth/js_web_bluetooth.dart';
 import 'package:mutex/mutex.dart';
 import 'package:rxdart/rxdart.dart';
 
-import 'BLELighthouseBackEnd.dart';
 import '../LighthouseDevice.dart';
 import '../adapterState/AdapterState.dart';
 import '../ble/Guid.dart';
+import 'BLELighthouseBackEnd.dart';
 import 'PairBackEnd.dart';
 import 'flutterWebBluetooth/FlutterWebBluetoothDevice.dart';